Fixed critical flaw in bag creation
From the BDBag specification:

> One or more (only one of each) of the following algorithm
>    (md5, sha1, sha256, sha512)
> Length of the file

BDBag successfully creates bags **Omitting all entries for which
the criteria are not met**. This happens silently and without any
kind of notification that the files have been omitted.

The following fix adds the empty string and zero length for each
entry missing those fields, so the files will be added successfully
on bag creation.