Added support for a wider range of Globus URLs
URLs in Globus can be specified a few different ways in bags and
manifests. If an endpoint and path can be derived out of a URL,
it will be allowed. Added support for the following:
    * https://<globus_endpoint>.data.globus.org/ (gcs v4)
    * https://<globus_endpoint>.e.globus.org/ (gcs v4 — petrel)
    * globus://<globus_endpoint>/

DID NOT add support for the following:
    * https://f8d036.eb38.dn.glob.us/ (gcs v5)

Not sure how to do GCS V5 yet...